consolidate tests
