+++ integration_test/test_compute_metrics.py
@@ -1,4 +1,10 @@
 import sigopt.xgboost
+from sigopt.xgboost.compute_metrics import (
+  compute_classification_report,
+  compute_mae,
+  compute_mse,
+  compute_accuracy
+)
 from sklearn.metrics import (
   accuracy_score,
   classification_report,
@@ -10,6 +16,21 @@ from sklearn.model_selection import train_test_split
 import xgboost as xgb
 import numpy as np
 
+
+def assert_classification_metrics(y_true, y_pred):
+  report_compute = compute_classification_report(y_true, y_pred)
+  report_sklearn = classification_report(y_true, y_pred, output_dict=True, zero_division=0)
+  assert np.abs(report_compute['weighted avg']['precision'] - report_sklearn['weighted avg']['precision']) < 1e-8
+  assert np.abs(report_compute['weighted avg']['recall'] - report_sklearn['weighted avg']['recall']) < 1e-8
+  assert np.abs(report_compute['weighted avg']['f1-score'] - report_sklearn['weighted avg']['f1-score']) < 1e-8
+  assert np.abs(compute_accuracy(y_true, y_pred) - accuracy_score(y_true, y_pred)) < 1e-8
+
+
+def assert_regression_metrics(y_true, y_pred):
+  assert np.abs(mean_absolute_error(y_true, y_pred) - compute_mae(y_true, y_pred)) < 1e-8
+  assert np.abs(mean_squared_error(y_true, y_pred) - compute_mse(y_true, y_pred)) < 1e-8
+
+
 iris = datasets.load_iris()
 X = iris.data
 y = iris.target
@@ -19,8 +40,8 @@ D_train = xgb.DMatrix(X_train, label=Y_train)
 D_test = xgb.DMatrix(X_test, label=Y_test)
 
 
-class TestComputeMetrics(object):
-  def test_compute_classification_metrics(self):
+class TestComputeXGBMetrics(object):
+  def test_classification_metrics(self):
     params = {
       'eta': 0.3,
       'num_class': num_class,
@@ -48,7 +69,7 @@ class TestComputeMetrics(object):
     assert np.abs(run.values['test0-F1'].value - report_sklearn['weighted avg']['f1-score']) < 1e-8
     assert np.abs(run.values['test0-accuracy'].value - accuracy_sklearn) < 1e-8
 
-  def test_compute_regression_metrics(self):
+  def test_regression_metrics(self):
     params = {
       'eta': 0.3,
       'objective': 'reg:squarederror'
@@ -72,3 +93,38 @@ class TestComputeMetrics(object):
     run = sigopt.get_run(ctx.run.id)
     assert np.abs(run.values['test0-mean absolute error'].value - mae) < (1e-8 / mae)
     assert np.abs(run.values['test0-mean squared error'].value - mse) < (1e-8 / mae)
+
+
+class TestComputeSyntheticMetrics(object):
+  def test_binary(self):
+    n = 50
+    y_true = np.random.randint(0, 2, n)
+    y_pred = np.random.randint(0, 2, n)
+    assert_classification_metrics(y_true, y_pred)
+
+  def test_multiclass(self):
+    n = 50
+    n_class = 3
+    y_true = np.random.randint(0, n_class, n)
+    y_pred = np.random.randint(0, n_class, n)
+    assert_classification_metrics(y_true, y_pred)
+
+  def test_zero_divide_multiclass(self):
+    n = 50
+    n_class = 3
+    y_true = np.random.randint(0, n_class, n)
+    y_pred = np.zeros(n)
+    assert_classification_metrics(y_true, y_pred)
+
+  def test_zero_divide_binary(self):
+    n = 50
+    y_true = np.random.randint(0, 2, n)
+    y_pred = np.zeros(n)
+    assert_classification_metrics(y_true, y_pred)
+
+  def test_regression_metrics(self):
+    # Check regression metrics
+    n = 100
+    y_true = np.random.randn(n)
+    y_pred = np.random.randn(n)
+    assert_regression_metrics(y_true, y_pred)
